refactor(audit): share byte formatting helper
