+++ src/lib/audit/workflow-log-summary.ts
@@ -4,29 +4,12 @@ import type {
   WorkflowRunStatus,
   WorkflowRunSummary,
 } from './workflow-log-types'
+import { formatBytes } from './format-bytes'
 import {
   getWorkflowEventStatus,
   getWorkflowEventTables,
 } from './workflow-log-helpers'
 import { WORKFLOW_LOG_PREFIX } from './workflow-log-types'
-
-const BYTE_UNITS = ['B', 'KB', 'MB', 'GB'] as const
-
-function formatBytes(value: number): string {
-  if (!Number.isFinite(value) || value <= 0) {
-    return '0 B'
-  }
-
-  let size = value
-  let unitIndex = 0
-  while (size >= 1024 && unitIndex < BYTE_UNITS.length - 1) {
-    size /= 1024
-    unitIndex++
-  }
-
-  const formatted = size >= 10 || unitIndex === 0 ? size.toFixed(0) : size.toFixed(1)
-  return `${formatted} ${BYTE_UNITS[unitIndex]}`
-}
 
 function getEventStatus(event: CacheLogEvent): WorkflowRunStatus {
   return getWorkflowEventStatus(event)
